+++ src/core/Collection.js
