+++ src/core/CollectionIndex.js
