+++ src/core/EnvironmentAdapter.js
