+++ src/core/StoreIndex.js
