+++ test/benchmarks/benchmark.update.js
@@ -50,7 +50,7 @@ ipfs.on('error', (err) => console.error(err))
 ipfs.on('ready', async () => {
     const run = async () => {
         try {
-            OrbitDB.addDatabaseType("aviondb.collection", require('../../src/Collection'))
+            OrbitDB.addDatabaseType("aviondb.collection", require('../../src/core/Collection'))
             const orbit = await OrbitDB.createInstance(ipfs, { directory: './orbitdb/benchmarks' })
 
             const db = await orbit.create('orbit-db.benchmark', "aviondb.collection", {
